+++ java/src/ie/tudublin/Main.java
@@ -1,6 +1,7 @@
 package ie.tudublin;
 
 import C21325616.MichaelsVisuals;
+import C22328351.LarinasVisual;
 import c22371846.*;
 import example.*;
 import C21325616.*;
@@ -13,7 +14,7 @@ public class Main {
 
     public void startUI() {
         String[] a = { "MAIN" };
-        processing.core.PApplet.runSketch(a, new Heartbeat());
+        processing.core.PApplet.runSketch(a, new PatricksVisuals());
     }
 
     public static void main(String[] args)
